### 2025-08-07 d948557
affiliation quoting and doi addition
