### 2025-08-09 14cb20a
Add with_params arg to get_spin_spiral_df
