### 2025-08-20 92f4ac1
Shorten the paper (~ 2000 words --> ~1200 words). Previous version in paper_old.md
