+++ Lumabi_JOSS/paper.md
@@ -29,7 +29,7 @@ authors:
     affiliation: "1"
 
 affiliations:
- - name:  Institute of Condensed Matter and Nanosciences, Université catholique de Louvain, B-1348 Louvain-la-Neuve, Belgium
+ - name:  Institute of Condensed Matter and Nanosciences, Université catholique de Louvain, B-1348 Louvain-la-Neuve, Belgium
    index: 1
  - name:  WEL Research Institute, avenue Pasteur 6, 1300 Wavre, Belgium.
    index: 2
@@ -42,78 +42,47 @@ affiliations:
 
 date: 07 August 2025
 bibliography: paper.bib
-
 ---
 
 # Summary
 
 ![Lumabi logo](Lumabi_logo.pdf){width=50%}
 
-Lumabi is a Python package integrated within the AbiPy framework [@gonze2020abinit] designed to automate and streamline the computation of phonon-resolved luminescence spectra of point defects (including the case of dopants) in inorganic solids using the ABINIT first-principles software application [@gonze2002first;@gonze2009abinit;@gonze2016recent;@gonze2020abinit]. The package addresses the growing need for efficient, reproducible workflows in materials science [@lejaeghere2016reproducibility;@bosoni2024verify], particularly in the study of defect-related luminescent properties, which are critical for applications ranging from quantum technologies [@wolfowicz2021quantum;@dreyer2018first] to down-conversion phosphors materials used in white light LEDs [@pust2015revolution;@lin2017inorganic;@fang2022evolutionary]. Lumabi automates key steps in the computational workflow, from initial $\Delta$SCF density-functional theory calculations with constrained occupations, to the generation of defect phonon modes in large supercells, right through the final generation of luminescence spectra based on the Huang-Rhys theory [@huang1950theory;@jin2021photoluminescence].
-A collection of tutorials and examples, organized as Jupyter Books,
-is available on the [Abipy Book project webpage](https://abinit.github.io/abipy_book/lumabi/intro/intro.html).
+Lumabi is a Python package within the AbiPy framework [@gonze2020abinit] that automates the computation of phonon-resolved luminescence spectra of point defects and dopants in inorganic solids using the ABINIT first-principles software [@gonze2020abinit]. The package provides an end-to-end workflow: from $\Delta$SCF density-functional theory calculations with constrained occupations, to the generation of defect phonon modes in large supercells, right through the generation of luminescence spectra based on Huang–Rhys theory [@huang1950theory;@jin2021photoluminescence].  
+
+Lumabi addresses the growing need for reproducible, automated workflows in defect physics [@lejaeghere2016reproducibility;@bosoni2024verify], with applications ranging from quantum technologies [@wolfowicz2021quantum;@dreyer2018first] to phosphors for solid-state lighting [@pust2015revolution;@lin2017inorganic;@fang2022evolutionary]. Tutorials and examples are available in the [AbiPy Book](https://abinit.github.io/abipy_book/lumabi/intro/intro.html).
 
 # Statement of need
 
-The study of defect-induced luminescence in materials is crucial for understanding and designing materials with specific optical properties. However, the computational workflow required to accurately predict these properties is complex and involves ground- and excited-state calculations,
-phonon computations, as well as pre- and post-processing steps.
+Defect-induced luminescence plays a key role in materials design for optoelectronics, quantum information, and phosphor technologies. Accurate predictions require ground- and excited-state calculations, phonon computations, and multiple post-processing steps, which are typically laborious to set up.  
 
[…]
