+++ Lumabi_JOSS/paper.bib
@@ -125,29 +125,6 @@
  year = {2002}
 }
 
-@article{gonze2009abinit,
- author = {Gonze, Xavier and Amadon, Bernard and Anglade, P-M and Beuken, J-M and Bottin, Fran{\c{c}}ois and Boulanger, Paul and Bruneval, Fabien and Caliste, Damien and Caracas, Razvan and C{\^o}t{\'e}, Michel and others},
- doi = {10.1016/j.cpc.2009.07.007},
- journal = {Computer Phys. Commun.},
- number = {12},
- pages = {2582--2615},
- publisher = {Elsevier},
- title = {ABINIT: First-principles approach to material and nanosystem properties},
- volume = {180},
- year = {2009}
-}
-
-@article{gonze2016recent,
- author = {Gonze, Xavier and Jollet, Fran{\c{c}}ois and Araujo, F Abreu and Adams, Donat and Amadon, Bernard and Applencourt, Thomas and Audouze, Christophe and Beuken, J-M and Bieder, Jordan and Bokhanchuk, A and others},
- doi = {10.1016/j.cpc.2016.04.003},
- journal = {Computer physics communications},
- pages = {106--131},
- publisher = {Elsevier},
- title = {Recent developments in the ABINIT software package},
- volume = {205},
- year = {2016}
-}
-
 @article{gonze2020abinit,
  author = {Xavier Gonze and Bernard Amadon and Gabriel Antonius and Frédéric Arnardi and Lucas Baguet and Jean-Michel Beuken and Jordan Bieder and François Bottin and Johann Bouchet and Eric Bousquet and Nils Brouwer and Fabien Bruneval and Guillaume Brunin and Théo Cavignac and Jean-Baptiste Charraud and Wei Chen and Michel Côté and Stefaan Cottenier and Jules Denier and Grégory Geneste and Philippe Ghosez and Matteo Giantomassi and Yannick Gillet and Olivier Gingras and Donald R. Hamann and Geoffroy Hautier and Xu He and Nicole Helbig and Natalie Holzwarth and Yongchao Jia and François Jollet and William Lafargue-Dit-Hauret and Kurt Lejaeghere and Miguel A.L. Marques and Alexandre Martin and Cyril Martins and Henrique P.C. Miranda and Francesco Naccarato and Kristin Persson and Guido Petretto and Valentin Planes and Yann Pouillon and Sergei Prokhorenko and Fabio Ricci and Gian-Marco Rignanese and Aldo H. Romero and Michael Marcus Schmitt and Marc Torrent and Michiel J. {van Setten} and Benoit {Van Troeye} and Matthieu J. Verstraete and Gilles Zérah and Josef W. Zwanziger},
  doi = {10.1016/j.cpc.2016.04.003},
