+++ Lumabi_JOSS/get_diff.sh
@@ -1,0 +1,4 @@
+# Get pandiff first with
+# npm install -g pandiff
+
+pandiff paper_old.md paper.md -o diff.pdf
