### 2025-08-20 9e603df
Fix GWR interpolation with SKW when using qprange
