+++ abipy/electrons/gwr.py
@@ -15,7 +15,6 @@ from functools import cached_property
 from monty.collections import dict2namedtuple
 from monty.string import list_strings, marquee
 from monty.termcolor import cprint
-#from abipy.core.func1d import Function1D
 from abipy.core.structure import Structure
 from abipy.core.mixins import AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter
 from abipy.core.kpoints import Kpoint, KpointList, Kpath, IrredZone, has_timrev_from_kptopt
@@ -610,6 +609,14 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
             keys = self.params.keys() if verbose else ["ecut", "ecutwfn", "ecutsigx", "ecuteps", "gwr_boxcutmin", "gwr_max_hwtene"]
             for k in keys:
                 app("%s: %s" % (k, self.params[k]))
+
+            app("Max_bstart: %d" % self.r.max_bstart)
+            app("min_bstop: %d" % self.r.min_bstop)
+            app("bstart_sk, bstop_sk")
+            for spin in range(self.nsppol):
+                for ikc, sigma_kpoint in enumerate(self.sigma_kpoints):
+                    app(f"{self.r.bstart_sk[spin, ikc]}, {self.r.bstop_sk[spin, ikc]}")
+
             if verbose:
                 app(marquee("k-points in Sigma_nk", mark="="))
                 app(self.r.sigma_kpoints.to_string(title=None, pre_string="\t"))
@@ -806,7 +813,7 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
                     None if ``ks_ebands_kmesh`` is not passed.
                 * interpolator: |SkwInterpolator| object.
         """
-        # TODO: Consistency check.
+        # Consistency check.
         errlines = []
         eapp = errlines.append
         if len(self.sigma_kpoints) != len(self.ebands.kpoints):
@@ -826,7 +833,6 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
 
         if ks_ebands_kpath is None:
             # Generate k-points for interpolation. Will interpolate all bands available in the GWR file.
-            bstart, bstop = 0, -1
             if vertices_names is None:
                 vertices_names = [(k.frac_coords, k.name) for k in self.structure.hsym_kpoints]
             kpath = Kpath.from_vertices_and_names(self.structure, vertices_names, line_density=line_density)
@@ -838,75 +844,84 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
             kfrac_coords = [k.frac_coords for k in ks_ebands_kpath.kpoints]
             knames = [k.name for k in ks_ebands_kpath.kpoints]
 
-            # Find the band range for the interpolation.
-            bstart, bstop = 0, ks_ebands_kpath.nband
-            bstop = min(bstop, self.r.min_bstop)
+            #bstop = min(bstop, self.r.min_bstop)
             if ks_ebands_kpath.nband < self.r.min_bstop:
                 cprint("Number of bands in KS band structure smaller than the number of bands in GW corrections", "red")
                 cprint("Highest GW bands will be ignored", "red")
 
             if not ks_ebands_kpath.kpoints.is_path:
-                cprint("Energies in ks_ebands_kpath should be along a k-path!", "red")
-
-            print(f"{ks_ebands_kpath.eigens.shape}=")
-            print("gwr_bstart", bstart)
[…]
