### 2025-08-20 7b6c962
update anaddb tests to force BEC calculation
