### 2025-08-20 9ed93ba
Revert "update anaddb tests to force BEC calculation"
