### 2025-08-21 f8b3432
Add dieflag only if it's not in anaddb_kwargs
