### 2025-09-04 #8728 @espottesmith
Hey @jbouquiaux, I'll serve as the editor for this submission.

Can you provide a list of possible reviewers for your submission? You can look at the related JOSS papers listed above or check out our [reviewer database](https://reviewers.joss.theoj.org/reviewers).
