### 2025-09-05 #8728 @yuan-gist
Hi @jbouquiaux , thanks for inviting. I would like to ask first, are your code inputs based on Abinit calculations?
