### 2025-09-08 #8728 @jbouquiaux
Hi @yuan-gist, the code is an extension of AbiPy which is designed to pre- and post-process Abinit calculations. As such, the workflow part that automates the DFT computations will run Abinit calculations.

That said, the other modules are more flexible and can work with results that are not Abinit-specific. For instance, a `DeltaSCF` object can indeed be constructed directly from Abinit output files,
```
    def from_four_points_file(cls, filepaths) -> DeltaSCF:
```
but it’s also possible to instantiate objects from data generated by other codes (though not directly from their raw output files). For the example above, one could initialize explicitly as follows:
```
    def __init__(self,structuregs,structureex,forces_gs,forces_ex,
                 ag_energy,ag_star_energy,ae_star_energy,ae_energy,meta=None):
```
Similarly, the Lineshape object does not rely on Abinit-specific outputs, since it’s constructed from more general arguments:
```
    def __init__(self, E_zpl, ph_eigvec, ph_eigfreq, structure,
                 forces,displacements,use_forces):
```
