### 2025-09-10 762163d
Add with_band_index to show band number of ebands.plot and phbands.plot
