### 2025-09-11 5fc7ab3
Print cut3d stderr
