### 2025-09-17 f336879
update pjdos plotting by type, better colormap mapping
