### 2025-10-06 #9145 @editorialbot
[REVIEW]: Lumabi: a Python package to streamline the computation of phonon-resolved luminescence spectra of defects and dopants in solids
**Submitting author:** @jbouquiaux (<a href="http://orcid.org/0000-0003-1982-052X">Julien Bouquiaux</a>)
**Repository:** https://github.com/abinit/abipy
**Branch with paper.md** (empty if default branch): lumabi_JOSS
**Version:** 0.9.8
