### 2025-10-08 568fee9
Make gstore.py compatible with new format
