### 2025-10-23 328eb1d
Replace get_number_of_atoms by len(supercell)
