+++ abipy/flowtk/abiphonopy.py
@@ -131,7 +131,7 @@ class PhonopyWork(Work):
 
         # Extract forces from the main Abinit output files.
         forces_filenames = [task.output_file.path for task in self.phonopy_tasks]
-        num_atoms = supercell.get_number_of_atoms()
+        num_atoms = len(supercell) #.get_number_of_atoms()
         force_sets = parse_set_of_forces(num_atoms, forces_filenames)
 
         # Write FORCE_SETS file.
