### 2025-10-24 ed5fe64
Report unconverged tasks properly
Check that we reached the maximum number of launches before trying to restart a task.
This allows to detect the unconverged status of a task.
Otherwise, the output file is renamed and the status is not correctly detected.
