### 2025-10-27 ee785f9
adding wfq factory
