### 2025-11-10 #9145 @jbouquiaux
Hi @alexsquires, you're right in your analysis. And since it's a [contribution to an exisiting package](https://joss.readthedocs.io/en/latest/submitting.html#should-i-write-my-own-software-or-contribute-to-an-existing-package), this seems to complicate the matter a bit.

Yes I did write the Lumabi functionality.
To help clarify the scope for review, this corresponds to the parts of the code described in the Abipy Book tutorial here:
https://abinit.github.io/abipy_book/lumabi/intro/intro.html

More specifically, the parts of the code up for review are:
- Workflow: abipy/flows/lumiwork.py
- Post-processing: abipy/lumi/**.py
- Embedding: abipy/embedding/embedding_ifc.py
