### 2025-12-03 6f2c816
Add support for frequency metadata in DDB file
