### 2025-12-03 355b4ba
update vars database, test python 3.14
