### 2025-12-04 02e0c89
Workaround to make phonopy_to_abinit work with new DDB format
