### 2025-12-04 36a3c81
Change ddb.header.lines as well
