### 2025-12-09 abinit/abipy#372 @seyedehsanhadi
Installation and Dependency Problems for JOSS Review #9145
1. Please kindly include the wannier90/ reference data directory in MANIFEST.in so tutorials execute without FileNotFoundError, or provide clear download instructions if obtained separately. Additionally, fix the embedding module import failure in conda-forge installations (works from GitHub source but not standard conda).

2. The tutorial lineshape_slider.py imports ipywidgets listed only in requirements-optional.txt, not mandatory dependencies. Please either move to mandatory dependencies or add try-except handling with clear user guidance.
