### 2025-12-09 abinit/abipy#373 @seyedehsanhadi
Convergence and Relaxation Validation for JOSS Review #9145
1. Workflows currently accept ABINIT inputs with zero convergence criteria, no toldfe, tolvrs, or tolwfr specified. This allows jobs to run with ABINIT defaults that may produce unconverged results. Please require at least ONE convergence criterion before proceeding and raise clear errors when violated.

2. The workflow accepts empty relax_kwargs dictionaries, ntime=0 (no ionic steps), missing tolmxf (uses loose defaults), and simultaneous toldfe AND tolvrs (conflicting). Please enforce: (a) non-empty relaxation parameters, (b) ntime > 0, (c) tolmxf specified or sensible default with warning, (d) reject or warn on multiple convergence criteria.

3. The DeltaSCF.from_four_points_file() method crashes with cryptic "list index out of range" for <4 files and silently ignores extras for >4 files. Kindly validate exactly 4 files upfront and raise clear ValueError: "Expected 4 GSR files (Ag, Ag*, Ae*, Ae), got N".

4. Validate ndivsm: reject if zero, enforce -50 ≤ ndivsm < 0 or ndivsm > 0, verify it's integer (not silently converted). Also add range checks for convergence tolerances: reject or warn when values fall outside reasonable ranges. Extreme values like toldfe=1.0 produce meaningless results silently.
