### 2025-12-09 abinit/abipy#374 @seyedehsanhadi
Memory, Threading, and Platform Compatibility for JOSS Review #9145
1. Break circular references between LumiWork, TaskManager, and Flow objects causing unbounded memory growth in loops. Additionally, add explicit thread synchronization across ALL TaskManager methods (currently only launch has locks).

2. Replace hard-coded /tmp and /dev/null with tempfile.gettempdir() and os.devnull. Enforce newline="\n" in AbinitInput.write() to prevent Windows CRLF issues. It's best to test cross-platform compatibility explicitly.
