### 2025-12-09 abinit/abipy#375 @seyedehsanhadi
Phonon Handling and Spectral Resolution for JOSS Review #9145
1. The hardcoded ±10 picosecond time window in utils_lumi.py imposes arbitrary ~0.2 meV spectral resolution and causes aliasing if high-frequency modes exceed Nyquist. Please make this adaptive based on maximum phonon frequency. Additionally, the Bose_einstein function lacks safeguards: zero or negative frequencies cause division by zero or nonsensical negative occupations.

2. The code silently treats imaginary frequencies (negative values in Phonopy, indicating structural instabilities) as valid modes or suppresses them without warning. For saddle point geometries, this produces meaningless spectra. Please add explicit detection of unstable structures and reject calculations or warn prominently. It's ideal to fail rather than silently produce wrong results.

3. The ASR enforcement in embedding_ifc.py uses diagonal-only correction without validating residual violations or checking for artificial unstable modes at boundaries. If embedding radius is too small, "blind" correction creates spurious modes. Please validate ASR residuals before correcting and check post-correction acoustic frequencies are physically reasonable (< 10⁻⁶ THz).
