### 2025-12-09 abinit/abipy#376 @seyedehsanhadi
Units, NAC, DDB, and Atom Mapping for JOSS Review #9145
1. Add validation that Phonopy objects use ABINIT conventions before applying fixed AMU→kg and eV/Å→N conversions in Delta_Q_nu(). VASP and other calculators use different conventions, corrupting results silently. Please check calculator type and validate eigenvector normalization or raise clear errors.

2. The hardcoded NAC factor 14.399652 in converters.py line 78 has explicit code comment admitting uncertainty about correctness. This affects LO-TO splitting in polar materials critically. Kindly either: (a) provide literature proof this is correct for ABINIT, (b) implement round-trip Abinit→Phonopy→Abinit validation preserving LO-TO splitting, or (c) document prominently that users must verify NAC for their materials.

3. The code forces DDB version downgrade (if ddb.version >= 20240201: ddb.version = 20230401) without explaining what data is discarded or validating critical properties (BEC, dielectric tensors, force constants) are preserved. This silently corrupts results from recent ABINIT versions. Please either: (a) support new DDB format, (b) add version checking refusing incompatible files with clear errors, or (c) document supported ABINIT versions with warnings.

4. The get_matching_dSCF_phonon_spcell function uses Cartesian distance checks ignoring periodic boundaries, so atoms crossing cell boundaries (0.99 → 0.01) cause silent mismatches and incorrect Huang-Rhys factors. Also, Lineshape.from_phonopy_phonons() uses hardcoded 0.3 Å tolerance, problematic for Jahn-Teller defects with large reconstructions. Please implement wrapped distance calculations accounting for periodicity and make tolerance adaptive to structural changes.
