### 2025-12-09 abinit/abipy#377 @seyedehsanhadi
Excited States, Bands, and BEC for JOSS Review #9145
1. The code silently falls back to HGH_TABLE pseudopotentials when none provided, making PAW/USPP calculations produce nonsense without warning. Please remove this fallback and require explicit pseudopotential specification instead.

2. The workflow lacks validation that converged charge density corresponds to intended excitation—no checks for magnetic moment, orbital character, or localization. Convergence to ground state or delocalized configurations produces irrelevant results silently. Please add validation examining converged state properties and warn or reject when excited state characteristics are questionable.

3. The default nb_extra_bands=10 for non-SCF calculations isn't validated for sufficiency, particularly problematic for f-electron systems like Eu²⁺ with high DOS where excited state manifolds may be truncated. Add checks examining local DOS and warning if truncation appears likely.

4. The code assigns BEC using heuristic stru_emb[-1].specie.common_oxidation_states[0], hazardous for transition metals and lanthanides (Eu²⁺ vs Eu³⁺) with multiple stable oxidation states depending on coordination. This produces incorrect LO-TO splitting. Please either: (a) provide quantitative comparison of oxidation-state approximation errors, (b) recommend only computed-BEC approach for quantitative work, or (c) add guidance on when to use which method.
