### 2025-12-09 abinit/abipy#378 @seyedehsanhadi
Data Portability and Internal Consistency for JOSS Review #9145
1. The workflow serializes results using absolute file paths (self.gs_relax_task.gsr_path), making output non-portable. Moving or archiving directories breaks subsequent post-processing. Please use relative paths or store data paths as configurable, allowing workflow relocation without breaking references. [optional]

2. The Lineshape class sums partial Huang-Rhys factors from phonon projections but never validates against global from real-space displacements in DeltaSCF. This comparison quantifies "missing" coupling from phonon basis incompleteness or projection errors. Please add consistency check comparing both approaches and warning if discrepancy exceeds ~5%.

3. The clean_structure function in utils_lumi.py reorders coordinates independently for DeltaSCF and Phonopy structures without ensuring atom indexing matches force constant matrices. This causes force constants of atom incorrectly applied to atom's displacement. Ensure cleaning maintains consistent indexing across structures or validate alignment afterward.
