### 2025-12-09 abinit/abipy#379 @seyedehsanhadi
Limitations, Convergence, and Physical Approximations for JOSS Review #9145
1. Please add dedicated "Limitations" section stating: (a) method calculates only radiative lineshape, not full quantum efficiency, (b) non-radiative recombination (SRH, Auger) excluded, (c) Franck-Condon approximation validity, (d) spin-orbit coupling not included (problematic for lanthanides), (e) default T=0 K assumption, (f) no breakdown criteria provided. Users must understand scope of applicability.

2. Provide quantitative recommendations for supercell sizes with systematic convergence evidence, k-point mesh adequacy, and convergence testing protocols. Include examples showing force convergence as function of supercell size. Currently users cannot systematically verify calculation quality.

3. The Poisson formula is mathematically invalid when ground and excited state frequencies differ. Since DeltaSCF calculates distinct eff_freq_gs and eff_freq_ex, ignoring this produces incorrect spectra for defects with mode softening. Please either derive frequency-dependent formula or document this limitation prominently. Correct the physics.

4. The calculation neglects change in zero-point energy (ZPE). Since the package calculates frequency changes between states, omitting ZPE introduces systematic error in predicted transition energy undocumented. Please include this term and explain its magnitude for typical defects.
