### 2025-12-09 abinit/abipy#380 @seyedehsanhadi
Mathematical Consistency and ASR Treatment for JOSS Review #9145
1. Equation 28 contains mathematical inconsistency: it requires quantities that appear as distinct elsewhere. FWHM formulas (Eqs. 41–42) introduce undefined notation disconnected from Equations 33–40. Please either define symbols consistently and show derivations, or remove disconnected formulas.

2. Equation 51 derives an expression that is undefined when frequency → 0 for acoustic modes at Γ. No discussion of acoustic treatment, frequency cutoffs, or regularization is provided. Provide rigorous proof acoustic modes are excluded from all summations, or show regularization procedure preserving translational invariance.

3. The diagonal-only ASR correction (Equation 53) lacks mathematical proof of sufficiency or literature citation validating this versus rigorous schemes. No quantitative assessment of residual ASR violations provided (acoustic frequencies < 10⁻⁶ THz?). Please provide theoretical justification or numerical evidence that residual errors are negligible compared to other approximation errors.
