### 2025-12-09 abinit/abipy#381 @seyedehsanhadi
Validation Evidence, Forces, and Error Analysis for JOSS Review #9145
1. The assumption that "vibrational modes in excited state are identical to ground state" lacks literature validation and quantitative breakdown criteria. Charge-transfer excitations substantially alter bonding and frequencies. Please either: (a) cite spectroscopy literature validating this for Eu²⁺ 4f→5d transitions in your hosts, (b) provide maximum allowable thresholds users should verify, or (c) flag as major assumption requiring case-by-case validation.

2. "forces are converged because they decay rapidly", circular reasoning not demonstrating what must be proven. Replace with explicit convergence evidence showing force decay versus supercell size (288-atom vs. 144-atom vs. 576-atom) with quantitative threshold (like, eV/Å).

3. The claim that "forces decay faster than displacements" is presented as universal but authors shows only one system with no comparison across defect types or theoretical justification. Please either cite literature establishing this as general principle or limit claim to a clearly defined scope, and add missing methodological details.

4. You mention Eu²⁺ 4f⁷ → 4f⁶5d¹ as excited state without spectroscopic justification or computed state ordering. Please add spectroscopic assignments or computed excited state energy ordering demonstrating appropriateness for your target materials.

5. The generating function introduces temperature dependence via Bose-Einstein factors without discussing physical validity of equilibrium statistical mechanics for non-equilibrium excited states. Please explain time scale hierarchy (vibrational thermalization vs radiative decay) with citations, to justify the framework.

6. The theory proceeds through a cascade of approximations (Franck-Condon, harmonic oscillators, identical modes, T=0 K, single-/multi-mode IFC) with no cumulative error analysis or ranking of error magnitudes. Add a section quantifying or at least ranking systematic errors so users understand which aspects require most careful validation.
