### 2025-12-09 #9145 @seyedehsanhadi
The authors have done an excellent job developing Lumabi, a well-timed and valuable contribution to computational materials research.
Before publication, a few improvements are needed. The software should include stronger input validation to prevent unrealistic or incorrect calculations. Some technical elements, such as platform-specific settings and unit handling, could be refined to ensure consistent and accurate results. Finally, the theoretical documentation should be expanded with clearer justification, references, and validation for key assumptions to help users understand the method’s limits and reliability.

To facilitate the revision process, I have grouped my findings into 10 thematic issues on the repository:

1. Installation & Dependencies: [https://github.com/abinit/abipy/issues/372#issue-3710983277]
2. Convergence & Relaxation Validation: [https://github.com/abinit/abipy/issues/373#issue-3711001365]
3. Memory, Threading & Platform Compatibility: [https://github.com/abinit/abipy/issues/374#issue-3711007363]
4. Phonon Handling & Spectral Resolution: [https://github.com/abinit/abipy/issues/375#issue-3711014871]
5. Units, NAC, DDB & Atom Mapping: [https://github.com/abinit/abipy/issues/376#issue-3711023734]
6. Excited States, Bands & BEC: [https://github.com/abinit/abipy/issues/377#issue-3711032265]
7. Data Portability & Internal Consistency: [https://github.com/abinit/abipy/issues/378#issue-3711043211]
8. Limitations, Convergence & Physical Approximations: [https://github.com/abinit/abipy/issues/379#issue-3711053803]
9. Mathematical Consistency & ASR Treatment: [https://github.com/abinit/abipy/issues/380#issue-3711065646]
10. Validation Evidence, Forces & Error Analysis: [https://github.com/abinit/abipy/issues/381#issue-3711093342]

I recommend addressing the validation and correctness issues first, as they directly impact the reliability of the results, followed by the documentation enhancements to clarify the physical assumptions and limitations for users. I will revisit the checklist items as authors address these issues and provide feedback on their responses.

*All tests were conducted in a WSL2 environment with the following setup:
-System: Linux 6.6.87.2-microsoft-standard-WSL2 (on Windows 11)
-Python version: 3.12.12
-Conda version: 25.9.1
