### 2025-12-25 74d1f46
Use decimal=4 when comparing e.elastic_relaxed to account to migration to codata2022
