### 2025-12-25 491a84f
set places to 2 for sv.sound_velocities
