+++ abipy/dfpt/tests/test_vsound.py
@@ -13,7 +13,7 @@ class SoundVelocityTest(AbipyTest):
         ddb_path = os.path.join(abidata.dirpath, "refs", "si_sound_vel", "Si_DDB")
         sv = SoundVelocity.from_ddb(ddb_path, num_points=5)
 
-        self.assertAlmostEqual(sv.sound_velocities[0,0], 5573.881381728927, places=4)
+        self.assertAlmostEqual(sv.sound_velocities[0,0], 5573.881381728927, places=2)
         self.assertEqual(sv.labels[0], "X")
 
         df = sv.get_dataframe()
