### 2025-12-26 a86b10d
Use decimal=4 in test_elastic
