### 2026-01-05 0048e04
Bug fix in make_doped_supercells, see https://github.com/openjournals/joss-reviews/issues/9145#issuecomment-3656810014
