### 2026-01-05 aae2bf2
Update MANIFEST.in to include additional file types and add embedding module for luminescence calculations.
