### 2026-01-05 7c39a22
Add validation for filepaths in DeltaSCF to ensure exactly 4 GSR files are provided
