### 2026-01-05 801dfab
Clarify ASR enforcement comment with reference to formalism in PhysRevMaterials
