### 2026-01-05 229ab83
Enhance ASR enforcement comment
