### 2026-01-09 #9145 @alexsquires
Apologies for the very slow review. I have two minor comments remaining, but otherwise I think Lumabi is a great package that will be well-used by the community. As they are small I have left them here rather than raised them as issues in the repo

- While there are tests for dedicated for Lumabi functionality within abipy, the scope seems quite limited. Could the authors explain the logic for the functionality that was selected for testing?

- within the paper, I noticed that one of the references is assigned incorrectly - Kavanagh et al., 2024 is cited as a luminsence post-processing package. It doesn't have this functionality. I couldn't find any other specific issues with the references in the paper, but if the authors could just check them, that would be great.
