### 2026-01-12 #9145 @jbouquiaux
> [@jbouquiaux](https://github.com/jbouquiaux) checking in (also happy New Year!)

Hi @espottesmith, happy new year!
I'm currently working on these three reviews.
