### 2026-01-15 0e6590b
Fix citation in paper.md
