### 2026-01-15 811120e
Compatibility with numpy 2.4
