### 2026-01-15 483488f
support read iterations in gwr
