### 2026-01-15 abinit/abipy#372 @jbouquiaux
Thank you for identifying these issues.

1. We have updated MANIFEST.in to include these data files.
The abipy/embedding directory was missing an __init__.py file, which caused import failures when the package is installed via conda-forge or pip. This worked from GitHub source due to how Python's path resolution works during development. We have added the missing __init__.py file.
aae2bf272dca8d8facc64d69a6e40c4fddb85d0d

2. Thank you for your attention to dependencies. However, lineshape_slider.py does not exist in the abipy repository.  It is part of the separate documentation repository (abipy_book), which has its own dependency management. For the abipy package itself, ipywidgets remains an optional dependency as intended, and is only required for interactive widget features.
