### 2026-01-15 abinit/abipy#373 @jbouquiaux
Regarding point 1, 2, and 4, we respectfully disagree. AbiPy is a workflow automation tool for users already familiar with ABINIT and DFT calculations. Users are responsible for providing appropriate input parameters.

Adding extensive input validation would duplicate ABINIT's own validation, impose arbitrary restrictions that may not apply to all use cases, and significantly increase maintenance burden. ABINIT itself validates inputs and provides error messages when parameters are invalid or conflicting.

For point 3, thank you for this suggestion. We have added (7c39a2239bf560b5d6bd9e11538294494c63baf1) validation to raise a clear error message if exactly 4 GSR files are not provided.
