### 2026-01-15 abinit/abipy#374 @jbouquiaux
1. We cannot locate the circular references or memory leaks you mention. Without precise file locations and test cases demonstrating the issue, we cannot reproduce or validate the problem.

2. We cannot locate the issues you mention. Hard-coded /tmp or /dev/null paths do not exist in the active codebase. Please provide specific file locations so we can investigate.
Regarding Windows CRLF: You mentionned that your tests were conducted in WSL2, which is a Linux environment where line endings are already Unix-style. ABINIT itself does not run natively on Windows and requires a Unix/Linux environment. Without a clear test case demonstrating the issue, we cannot validate the problem.
