### 2026-01-15 abinit/abipy#375 @jbouquiaux
1. The 10 picosecond time window (line 48 of utils_lumi.py) is appropriate for phonon spectroscopy in luminescence applications. This value was validated through extensive testing and is suitable for the typical phonon energy ranges encountered in these calculations. If you have identified specific cases where this resolution is insufficient or causes aliasing issues, please provide the details so we can investigate. Regarding the Bose_einstein function: it correctly handles T=0 (line 28-29). Regarding negative frequencies, kindly see the next point.

2. This is incorrect. The code does handle low and negative frequencies. In abipy/lumi/lineshape.py line 151, there is an explicit check: `if self.ph_eigfreq[i] < 1e-5`, which sets the contribution to zero for modes with frequencies below the threshold.

3. Thank you for this observation. The ASR enforcement (lines 207-213 of embedding_ifc.py) uses a diagonal-only correction following the formalism developed in https://doi.org/10.1103/PhysRevMaterials.5.084603 and https://doi.org/10.1103/PhysRevB.104.045303. Please see the discussions in these references for additional details on the method and its validation. We have clarified the code by adding a comment referencing these papers and added them to the jupyter book documentation (229ab834c4834deb0c0a602b8458020d4bc66293, 801dfab5f4d9f9769624c472866c4c809510b004, https://github.com/abinit/abipy_book/commit/3af663458e0f9e38e7d93f7e0b175f3d6cc785be#diff-967eed69ed5b024931b3740e1167534b324e438a204fd6c51e224722005539b3R398) . Users are responsible for choosing an appropriate embedding radius for their system and validating convergence, as optimal parameters are system-dependent and cannot be standardized across all defect types.
