### 2026-01-15 abinit/abipy#376 @jbouquiaux
1. The AMU-kg and eV/A-N conversions in Delta_Q_nu() are standard SI unit conversions, not ABINIT-specific conventions. Phonopy is calculator-agnostic by design. The from_phonopy_phonons() method accepts any Phonopy object regardless of source calculator. The eigenvectors and frequencies are already in Phonopy's standardized format. No calculator-specific validation is needed.
2. The NAC factor is empirically validated through round-trip testing (test_phonopy_ddb_phonopy in test_converters.py verifies Phonopy-ABINIT-Phonopy preserves BECs/dielectric tensors to 5 decimal places).
3. The reason for this downgrade is explained in the code.
    /# DDB version > 20240201 contains an additional line with the frequency of the perturbation
    /# This feature is not not supported in add_data_ddb and ddb.write.
    /# To bypass this problem, here we write the DDB using an old format without frequencies.
This downgrade is only there for the conversion and does not affect the subsequent results.
4. No periodic wrapping (0.99-0.01) should be present. `get_matching_dSCF_phonon_spcell` operates on supercells in Cartesian coordinates where all atoms have explicit positions without wraparound. The tolerance parameter exists because phonon calculations may use pristine atomic positions while $\Delta$-SCF calculations use defect-relaxed positions. The tolerance accommodates these structural differences between the two calculations. It's a tunable function parameter (not hardcoded), and the function prints clear warnings when matching fails, allowing users to adjust the tolerance for their specific defect geometry.
