### 2026-01-15 abinit/abipy#377 @jbouquiaux
1. Thank you for this comment. We note that this concern is outside the scope of the Lumabi paper, as it relates to AbiPy's general workflow infrastructure rather than luminescence functionality. Additionally, HGH table pseudopotentials are norm-conserving, not PAW/USPP. If you have concerns about this general AbiPy behavior, please open a separate issue.
2. We appreciate this suggestion. However, validation criteria for charge density characteristics (magnetic moments, orbital character, localization) are highly system-specific and vary significantly between materials and excitation types. Automated checks would likely produce false positives or negatives. It is hence the responsability of the user to check convergence.
3. Thank you for this concern. The choice of `nb_extra_bands` is system-dependent, and 10 may be insufficient for your systems. AbiPy provides visualization tools (https://abinit.github.io/abipy/gallery/index.html) for users to verify band sufficiency.
4. Thank you for this observation. The oxidation state heuristic is applied only to interstitial sites. This is a simple pragmatic choice. Validation tests showed that phonon frequencies are minimally affected by varying oxidation state assignments for interstitials.
