### 2026-01-15 abinit/abipy#378 @jbouquiaux
1. Thank you for this suggestion. The use of absolute paths is intentional for the workflow's runtime operation. For archival purposes, users can process the results before relocating directories.
2. This is indeed a good remark. You can compare the two approaches by using `S_em()` from the deltaSCF class (1D model, real displacements) and `S_tot()` from the lineshape class (multi-D model, sum over partial HR factors). Some discussion about this comparison is provided in https://doi.org/10.1002/adom.202100649.
3. the `clean_structure()` function (in `utils_ifc.py`, not in `utils_lumi.py`) does not reorder atoms. It only translates and adjusts coordinates while preserving indices.
