### 2026-01-15 abinit/abipy#379 @jbouquiaux
1. Limitations of the formalism are discussed in the tutorial's "Theory" section and referenced papers. Regarding your specific points: (d) spin-orbit coupling can be included via ABINIT (https://docs.abinit.org/tutorial/spin/), and (e) temperature is a tunable parameter in all relevant functions.
2. Convergence testing methods and examples are provided in the AbiPy gallery: https://abinit.github.io/abipy/gallery/index.html.
3. The Poisson formula assumes identical ground and excited state phonon frequencies (the "equal mode" approximation), as stated in the documentation: "We assume first that the vibrational modes in the excited electronic state are identical to those in the ground state." The mathematical derivation is provided in the referenced literature.
4. The change in ZPE due to phonon frequency differences can be computed using methods provided in the package. This term is typically negligibly small.
