### 2026-01-15 abinit/abipy#380 @jbouquiaux
1. Could you specify the notation inconsistency? We believe symbols are defined consistently and references are provided for derivations.
2. Kindly see https://github.com/abinit/abipy/issues/375, point 2.
3. Kindly see https://github.com/abinit/abipy/issues/375, point 3.
