### 2026-01-15 abinit/abipy#381 @jbouquiaux
1. This assumption requires indeed case-by-case validation. For information, some literature validation can be found in https://doi.org/10.1103/PhysRevMaterials.5.084603 (referenced in the paper).
2. We cannot locate this exact sentence in the paper or in the documentation.
3. Kindly see reference 10.1088/1367-2630/16/7/073026 (mentionned in the main text) for the methodological details.
4. Eu²⁺ 4f⁶5d¹ is the well-known excited state for Eu²⁺ luminescence (Shionoya et al., 2018, Phosphor Handbook). Additionally, we believe  that this software review should not focus on spectroscopic details.
5. These formalism details are covered in the referenced literature and are beyond the scope of this software paper.
6. Error analysis is material-specific and should be performed on a case-by-case basis.
