### 2026-01-16 abinit/abipy#372 @seyedehsanhadi
All three issues are satisfactorily addressed. The installation and dependency problems are resolved for JOSS review #9145.
