### 2026-01-16 abinit/abipy#375 @seyedehsanhadi
RESPONSE 3: ✅ ACCEPTED
RESPONSE 1:  ✅ ACCEPTED
RESPONSE 2: ❌ your response does not refute the criticism; it confirms that.

Your code treats imaginary frequencies (freq < 0, indicating dynamically
unstable structure) identically to acoustic modes (freq ≈ 0, expected at
Γ-point) with the same generic "ignored" warning. When a
user has an unrelaxed excited state with imaginary phonon modes, your code
continues the calculation and returns a luminescence spectrum that is
scientifically meaningless because it's computed from a saddle-point geometry,
not a stable minimum. Imaginary frequencies indicate the structure
needs further relaxation, this is a critical error that should STOP the
calculation, not proceed silently.

I suggest :

**Replace the entire `Delta_Q_nu` function with this corrected version:**

```python
def Delta_Q_nu(self):
    """
    List of Delta_Q_nu of the atoms in the structure, in SI unit
    """
    Q_nu = np.zeros(self.n_modes())

    masses = np.repeat(self.mass_list(), 3) * 1.66053892173E-27
    ph_eigvector = self.ph_eigvec
    ph_eigfreq = self.ph_eigfreq * (abu.eV_s)

    if self.use_forces:
        delta_forces = self.forces.flatten() * ((abu.eV_Ha)*(abu.Ha_J)/(1e-10))
        for i in range(self.n_modes()):
            if self.ph_eigfreq[i] < 0:
                raise ValueError(
                    f"Imaginary frequency detected: mode {i}, freq={self.ph_eigfreq[i]:.4e} eV. "
                    f"Structure is dynamically unstable (saddle point, not minimum). "
                    f"Action required: Verify ground/excited state geometry is fully relaxed."
                )
            elif self.ph_eigfreq[i] < 1e-5:
                warnings.warn(f"Near-zero mode {i} ({self.ph_eigfreq[i]:.2e} eV) set to zero (expected acoustic mode at Γ-point).")
                Q_nu[i] = 0
            else:
                Q_nu[i] = (1/ph_eigfreq[i]**2) * np.sum( delta_forces/np.sqrt(masses) * np.real(ph_eigvector[i]))
    else:
        displacements = self.displacements.flatten() * (1e-10)
        for i in range(self.n_modes()):
            if self.ph_eigfreq[i] < 0:
                raise ValueError(
                    f"Imaginary frequency detected: mode {i}, freq={self.ph_eigfreq[i]:.4e} eV. "
                    f"Structure is dynamically unstable (saddle point, not minimum). "
                    f"Action required: Verify ground/excited state geometry is fully relaxed."
                )
            elif self.ph_eigfreq[i] < 1e-5:
                warnings.warn(f"Near-zero mode {i} ({self.ph_eigfreq[i]:.2e} eV) set to zero (expected acoustic mode at Γ-point).")
                Q_nu[i] = 0
            else:
                Q_nu[i] = np.sum(np.sqrt(masses) * displacements * np.real(ph_eigvector[i]))

    Q_nu[0:3] = 0
    return Q_nu
```
