### 2026-01-16 abinit/abipy#373 @seyedehsanhadi
POINT 3: ✓ ACCEPTED

POINTS 1, 2, 4: ⚠️

While I respect your opinion, since validation was declined, users must be clearly warned that these inputs are not checked and that invalid inputs may lead to cryptic errors or silent failures; users are responsible for ensuring parameter validity.
