### 2026-01-16 abinit/abipy#374 @seyedehsanhadi
Thank you for your response.

**Windows CRLF Problem** ⚠️ **REQUIRES FIX**
Found in  methods in `abipy/abio/inputs.py`:
Current: `with open(filepath, "wt") as fh:`
Fix: `with open(filepath, "wt", newline="\n") as fh:`

Without `newline="\n"`, Windows converts LF to CRLF, breaking ABINIT input files.

Fix is trivial (add `newline="\n"` to 3 lines in `abipy/abio/inputs.py`). This ensures cross-platform compatibility without affecting Unix/Linux users. Other claims withdrawn.
