### 2026-01-16 abinit/abipy#376 @seyedehsanhadi
Thank you for the detailed responses. I agree with your explanations for issues 1, 3, and 4.
**Regarding issue 2 (NAC factor):**

Your code comment at converters.py lines 77-78 states:
```python
# according to the phonopy website 14.399652 is not the coefficient for abinit
# probably it relies on the other conventions in the output.
```

This contradicts the claim that the factor has been "empirically validated through round-trip testing."

The existing test_phonopy_ddb_phonopy verifies that BEC/dielectric tensors are preserved to 5 decimals, but this only confirms mathematical consistency, not that the NAC factor is physically correct for ABINIT. A wrong factor can preserve tensor values while still producing incorrect LO-TO splitting in phonon frequencies.

**Minimum request:** Please add a disclaimer in the code or documentation noting that users working with polar materials should independently validate LO-TO splitting, or remove the uncertainty comment if you're confident the factor is correct.
