### 2026-01-16 abinit/abipy#377 @seyedehsanhadi
All points resolved.
