### 2026-01-16 abinit/abipy#378 @seyedehsanhadi
All technical concerns addressed or refuted.
