### 2026-01-16 abinit/abipy#379 @seyedehsanhadi
✅ **1:** Acceptable
✅ **2:** Acceptable
⚠️ **3:** Documentation enhancement needed
⚠️ **4:** Scope clarification helpful, ZPE shift is zero under the equal‑mode approximation by definition; when this approximation breaks, the ZPE contribution is usually small compared to the optical transition energy, but its size is system‑dependent and should be documented: Documentation enhancement needed.
