### 2026-01-16 abinit/abipy#380 @seyedehsanhadi
For the purpose of this JOSS review, I consider these requests as documentation/clarity improvements rather than blocking issues, so I am happy to proceed and will close this issue after posting this comment.
