### 2026-01-16 #9145 @seyedehsanhadi
> Thank you for your review and for recognizing Lumabi as a valuable contribution to computational materials research. We have carefully examined each of the 10 thematic issues you raised and provided detailed point-by-point responses in the corresponding GitHub issues.
>
> We have made improvements where your comments identified legitimate areas for enhancement, particularly in clarifying documentation and strengthening certain validation checks. However, we must respectfully note that several comments in the review appear to be based on factual inaccuracies or misunderstandings of our implementation. Specifically:
>
> * **Factual errors about the codebase**: Multiple claims reference code behavior that does not exist in our implementation (e.g., "hardcoded 0.3 Å tolerance" in [Units, NAC, DDB, and Atom Mapping for JOSS Review #9145 abinit/abipy#376](https://github.com/abinit/abipy/issues/376), which is actually a configurable function parameter; claims about Phonopy using "ABINIT conventions" in issue [Units, NAC, DDB, and Atom Mapping for JOSS Review #9145 abinit/abipy#376](https://github.com/abinit/abipy/issues/376), which is factually incorrect; [Units, NAC, DDB, and Atom Mapping for JOSS Review #9145 abinit/abipy#376](https://github.com/abinit/abipy/issues/376) claims `clean_structure()` "reorders coordinates," when it only translates coordinates while preserving atom indexing.).
> * **Quoted text that does not exist**: [Units, NAC, DDB, and Atom Mapping for JOSS Review #9145 abinit/abipy#376](https://github.com/abinit/abipy/issues/376) references statements like "forces are converged because they decay rapidly" that we cannot locate in our code, paper, or documentation.
> * **Out-of-scope physics discussions**: Several comments request detailed physics validation and theoretical justification more appropriate for a physics journal than a JOSS software review, which focuses on software quality, documentation, and usability.
>
> We have addressed these concerns transparently in our individual responses, providing code references and clarifications where appropriate. We are happy to discuss any specific technical concerns further and to make additional improvements where they genuinely enhance the software.

Dear @jbouquiaux,
Thank you for the responses. I have reviewed the code references again and, unfortunately, there are still outstanding technical inconsistencies that conflict with the claims made in your rebuttal. These require resolution before I can recommend acceptance. For your note, I've closed issues 372, 377, 378, 380, and 381 as satisfactorily addressed.

Issue 375:
Your code at lineshape.py:151 uses if self.ph_eigfreq[i] < 1e-5. This treats imaginary phonon modes (freq < 0 = unstable structure) identically to acoustic modes (freq ≈ 0 = valid physics). Users unknowingly compute meaningless spectra from saddle-point geometries.

Issue 376:
Your code comment at converters.py states:
"# according to the phonopy website 14.399652 is not the coefficient for abinit"
You then claim this factor is "empirically validated". Both cannot be true. Your round-trip test validates mathematical consistency, NOT physical correctness of LO-TO splitting.

Issue 374:
Exact location provided: abipy/abio/inputs.py. Trivial fix.
​
Issue 379:
Points 3 & 4 require documentation enhancements per my evaluation. JOSS requires clear limitation statements. Equal-mode approximation breakdown and ZPE contributions must be documented.
​
A kind note on "Out-of-Scope Physics":
Documenting physics approximation validity, including breakdown regimes and magnitudes of neglected terms, is software documentation that must reside in the repository's technical documentation (README, API docs, or similar). JOSS explicitly requires this as part of API documentation in its review criteria, distinguishing it from research contributions. Such documentation should not appear in the JOSS paper itself, but reviewers are expected to verify its completeness in the software repository.

@espottesmith
I can recommend the manuscript for publication once these specific discrepancies between the code behavior and the documentation are resolved.

 KERNEL: 6.6.87.2-microsoft-standard-WSL2,  OS: Ubuntu 24.04.3 LTS,   CONDA: 25.9.1, Py: 3.13.9
