### 2026-01-19 02c056d
Add SCF iteration and plotting functionality in GwrFile class
