+++ abipy/electrons/gwr.py
@@ -469,6 +469,11 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
     def completed(self) -> bool:
         """True if GWR calculation completed."""
         return bool(self.r.read_value("gwr_completed", default=1))
+    
+    @cached_property
+    def scf_iteration(self) -> int:
+        """The number of SCF iterations performed in the self-consistent calculation."""
+        return int(self.r.read_value("scf_iteration", default=1))
 
     @property
     def sigma_kpoints(self) -> KpointList:
@@ -492,6 +497,14 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
         Shape: [nsppol, nkcalc]
         """
         return self.r.read_value("qpz_gaps") * abu.Ha_eV
+
+    # @cached_property
+    # def qp_pade_dirgaps(self) -> np.ndarray:
+    #     """
+    #     QP direct gaps in eV computed with the quasi-particle equation
+    #     Shape: [nsppol, nkcalc]
+    #     """
+    #     return self.r.read_value("qp_pade_gaps") * abu.Ha_eV
 
     @cached_property
     def minimax_mesh(self) -> MinimaxMesh:
@@ -635,7 +648,8 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
                               kpoint: KptSelect | None = None,
                               spin: int | None = None,
                               with_params: bool = True,
-                              with_geo: bool = False) -> pd.DataFrame:
+                              with_geo: bool = False,
+                              iter: bool = False) -> pd.DataFrame:
         """
         Return a pandas DataFrame with the QP direct gaps in eV.
 
@@ -645,20 +659,31 @@ class GwrFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
             spin: Spin index. None, to select all spins.
             with_params: True if GWR parameters should be included.
             with_geo: True if geometry info should be included.
+            iter: If True include data for all SCF iterations.
         """
         d = {}
-        d["kpoint"] = [k.frac_coords for k in self.sigma_kpoints] * self.nsppol
-        d["kname"] = [k.name for k in self.sigma_kpoints] * self.nsppol
-        d["ks_dirgaps"] = self.ks_dirgaps.ravel()
-        d["qpz0_dirgaps"] = self.qpz0_dirgaps.ravel()
-        #d["qp_pade_dirgaps"] = self.qp_pade_dirgaps.ravel()
-        d["spin"] = [0] * len(self.sigma_kpoints)
-        if self.nsppol == 2: d["spin"].extend([1] * len(self.sigma_kpoints))
+        d["kpoint"] = [k.frac_coords for k in self.sigma_kpoints] * self.nsppol * self.scf_iteration
+        d["kname"] = [k.name for k in self.sigma_kpoints] * self.nsppol * self.scf_iteration
+        if iter:
+            d["iteration"] = []
+            d["ks_dirgaps"] = np.zeros(0)
+            d["qpz0_dirgaps"] = np.zeros(0)
+            # d["qp_pade_dirgaps"] = np.zeros(0)
+            for iter in range(self.scf_iteration):
+                d["iteration"].extend([iter + 1] * len(self.sigma_kpoints) * self.nsppol)
+                d["ks_dirgaps"] = np.concatenate((d["ks_dirgaps"], (self.r.read_value("ks_gaps", path=f"iter{iter+1}") * abu.Ha_eV).ravel()))
[…]
