### 2026-01-20 dc7697e
add KS point in figure
