### 2026-01-23 3a22f04
Fix parsing of pseudos in AbinitInput
