### 2026-02-11 850121c
Fix tests with pandas3
