+++ abipy/ppcodes/oncv_plotter.py
@@ -171,13 +171,15 @@ class OncvPlotter(NotebookWriter):
             if what == "bound_states":
                 # Show position of the last peak.
                 s, marker = 10, "^"
+                #style = dict(color=self.color_l[l], s=s, marker=marker)
                 ae_peaks = ae_wf.get_peaks()
-                style = dict(color=self.color_l[l], s=s, marker=marker)
+                if len(ae_peaks.xs):
+                    ax.scatter(ae_peaks.xs[-1], ae_peaks.ys[-1], color=self.color_l[l])
 
-                ax.scatter(ae_peaks.xs[-1], ae_peaks.ys[-1], color=self.color_l[l])
                 ps_peaks = ps_wf.get_peaks()
-                style = dict(color=self.color_l[l], s=s, marker=marker)
-                ax.scatter(ps_peaks.xs[-1], ps_peaks.ys[-1], color=self.color_l[l])
+                #style = dict(color=self.color_l[l], s=s, marker=marker)
+                if len(ps_peaks.xs):
+                    ax.scatter(ps_peaks.xs[-1], ps_peaks.ys[-1], color=self.color_l[l])
 
             ax.plot(ae_wf.rmesh, ae_wf.values, label=fr"AE {nlk.latex}", **self._mpl_opts_laeps(l, "ae"))
             ax.plot(ps_wf.rmesh, ps_wf.values, label=fr"PS {nlk.latex}", **self._mpl_opts_laeps(l, "ps"))
