+++ abipy/abio/abivar_database/variables_abinit.py
@@ -2979,14 +2979,17 @@ Variable(
     abivarname="dmft_charge_prec",
     varset="dmft",
     vartype="real",
-    topics=['DMFT_expert'],
+    topics=['DMFT_expert', 'DmftTriqsCthyb_expert'],
     dimensions="scalar",
     defaultval=1e-06,
-    mnemonics="Dynamical Mean Field Theory: charge density precision",
+    mnemonics="Dynamical Mean Field Theory: CHARGE density PRECision",
+    requires="[[usedmft]] == 1",
     added_in_version="before_v9",
     text=r"""
-Precision to achieve in determining the charge density in the computation of the fermi level.
-Should be decreased to increase precision. However, for a large system, it can increase importantly computer time.
+This variable sets the numerical tolerance used to determine the Fermi level. The root-finding algorithm stops once the calculated electron count matches the target value within this tolerance.
+A smaller value improves the accuracy of the Fermi level but increases the number of root-finding iterations, which may significantly impact performance for large systems. If you are using noisy solvers (e.g., CT-QMC), it is unnecessary to set a tolerance smaller than the intrinsic noise level of your solver.
+
+See [[dmft_fermi_step]] for further tuning of the root-finding algorithm.
 """,
 ),
 
@@ -2994,20 +2997,33 @@ Variable(
     abivarname="dmft_dc",
     varset="dmft",
     vartype="integer",
-    topics=['DMFT_useful'],
+    topics=['DMFT_basic', 'DmftTriqsCthyb_compulsory'],
     dimensions="scalar",
-    defaultval=1,
+    defaultval=r"None if [[dmft_solv]] $\in$ [6,7], 1 otherwise",
     mnemonics="Dynamical Mean Field Theory: Double Counting",
-    characteristics=['[[DEVELOP]]'],
+    requires="[[usedmft]] == 1",
     added_in_version="before_v9",
     text=r"""
 
-Value of double counting used for DMFT (so, only relevant for [[usedmft]]=1)..
+Selects the double counting (DC) correction formula used in DFT+DMFT calculations.
 
-   * 1: corresponds to the "Full Localized Limit" double counting (to be used with [[usepawu]]=10).
-   * 2: corresponds to the "Around Mean Field" double counting (this is not yet in production).
-   * 5: the calculation is done without magnetism in the J term (cf [[cite:Park2015]] and [[cite:Chen2016a]]), to be used with [[usepawu]]=14.
-   * 6: this option is in development.
+   * 1 or 5 - Full Localized Limit (FLL): magnetic (1) and non-magnetic (5) versions.
+   * 2 or 6 - Around Mean Field (AMF): magnetic (2) and non-magnetic (6) versions.
+   * 7 - Nominal double counting: non-magnetic version only. Uses the nominal occupancy set by [[dmft_nominal]].
+   * 8 - Exact formula (cf [[cite:Haule2015a]]): non-magnetic version only, and only compatible
+         with LDA and PBE ([[ixc]] $\in$ [7,-1012,11,-101130]).
+         This is the formula you should always use. The implementation makes the assumption
+         that the screened potential has the form of a Yukawa potential (cf [[dmft_yukawa_param]]),
+         which is rigorously valid only in the case [[dmft_solv]]=7, with the full Slater Hamiltonian.
+         Besides, we assume that the projection of the correlated orbital [[dmft_orbital]]
[…]
