### 2026-02-13 f69420a
Fix docstrings and spacing
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
