+++ abipy/abio/inputs.py
@@ -2321,15 +2321,23 @@ with the Abinit version you are using? Please contact the AbiPy developers.""" %
         return inp
 
 
-    def make_phpert_input(self, perturbation, tolerance = None, prtwf = -1, prepgkk = 0, manager=None) -> AbinitInput:
+    def make_phpert_input(self, perturbation, tolerance=None, prtwf=-1, prepgkk=0, manager=None) -> AbinitInput:
         """
         Return |AbinitInput| for a Phonon calculation for a given perturbation.
         This functions should be called with an input that represents a GS run.
 
         Args:
-            perturbation: dict with the Abinit variables defining the irreducible perturbation
-                Example: {'idir': 1, 'ipert': 4, 'qpt': [0.0, 0.0, 0.0]},
-            manager: |TaskManager| of the task. If None, the manager is initialized from the config file.
+            perturbation: dict with the Abinit variables defining the irreducible perturbation.
+                Example: {'idir': 1, 'ipert': 4, 'qpt': [0.0, 0.0, 0.0]}.
+            tolerance: Dict with a single ABINIT tolerance variable (e.g. ``{'tolvrs': 1.0e-10}``)
+                used to control the convergence of the DFPT calculation. If ``None``, a
+                default of ``{'tolvrs': 1.0e-10}`` is used.
+            prtwf: Value passed to the ABINIT variable ``prtwf`` to control the printing
+                of wavefunctions for the phonon calculation. Defaults to ``-1`` (ABINIT default).
+            prepgkk: Value for the ABINIT variable ``prepgkk`` controlling the preparation
+                of GKK-related data for subsequent calculations. Defaults to ``0``.
+            manager: |TaskManager| of the task. If None, the manager is initialized from
+                the config file.
         """
         if tolerance is None: tolerance = {"tolvrs": 1.0e-10}
 
