### 2026-02-13 e089040
Fix docstrings and spacing
