### 2026-02-24 #9145 @jbouquiaux
Hi @espottesmith, I've completed the tasks mentioned in the list above.
Software version: 0.9.8
Archive :  [10.5281/zenodo.18756677](https://doi.org/10.5281/zenodo.18756677)

Please let me know if anything is missing.
