### 2026-02-25 #9145 @espottesmith
Paper comments:

- The plots within your figures (e.g., "luminescence lineshape" in Figure 3) are unreadable. Is it possible to increase the font sizes so that these plots can be read?
- Figures should be referenced in the text, where appropriate.
- A brief description of the "Huang-Rhys spectral function" and its physical meaning would be appreciated, though I recognize that you cite relevant works.

If you can address these points, I'll be happy to recommend acceptance!
