+++ Lumabi_JOSS/paper.md
@@ -72,10 +72,9 @@ each individual building block can be also used independently with ease. In this
 
 ## LumiWork Module
 
-![The LumiWork module, an AbiPy Workflow that automates ABINIT DFT tasks with $\Delta$SCF constrained occupations.](LumiWork.pdf)
+![The LumiWork module, an AbiPy Workflow that automates ABINIT DFT tasks with $\Delta$SCF constrained occupations.[]{label="fig-lumiwork"}](LumiWork.pdf)
 
-A computational workflow for calculating phonon-resolved photoluminescence (PL) spectra of defect systems starts with the LumiWork module, which automates ABINIT DFT tasks with $\Delta$SCF constrained occupations [@jones1989density;@hellman2004potential]. Users provide the defect supercell structure (e.g. in CIF format), the DFT input parameters (e.g., the plane-wave kinetic energy cutoff, the stopping SCF criterion, etc.), and constrained occupations of the Kohn-Sham states designed to mimic the excited state of the system under study. This module manages two structural relaxations for
-the ground- and the excited-state, and offers optional static SCF computations followed by non-SCF band structure calculations. Note that the workflow is dynamic: since the relaxed excited state is not known beforehand, input files must be generated at run-time.
+A computational workflow for calculating phonon-resolved photoluminescence (PL) spectra of defect systems starts with the LumiWork module (\autoref{fig-lumiwork}), which automates ABINIT DFT tasks with $\Delta$SCF constrained occupations [@jones1989density;@hellman2004potential]. Users provide the defect supercell structure (e.g. in CIF format), the DFT input parameters (e.g., the plane-wave kinetic energy cutoff, the stopping SCF criterion, etc.), and constrained occupations of the Kohn-Sham states designed to mimic the excited state of the system under study. This module manages two structural relaxations for the ground- and the excited-state, and offers optional static SCF computations followed by non-SCF band structure calculations. Note that the workflow is dynamic: since the relaxed excited state is not known beforehand, input files must be generated at run-time.
 
 The `LumiWork` class inherits from the Abipy's `Work` class.
 Consequently, it leverages error detection and automatic restart handling capabilities available for all ABINIT-related Works,
@@ -84,9 +83,9 @@ The majority of outputs are in NetCDF format, serving as the foundation for subs
 
 ## $\Delta$SCF Post-Processing Module
 
-![The $\Delta$SCF module, designed to post-process $\Delta$SCF constrained-occupation calculations using a one-dimensional configuration-coordinate model.](dSCF_post_process.pdf)
+![The $\Delta$SCF module, designed to post-process $\Delta$SCF constrained-occupation calculations using a one-dimensional configuration-coordinate model.[]{label="fig-post_process"}](dSCF_post_process.pdf)
 
-The next step in the workflow is handled by the $\Delta$SCF post-processing module. This tool takes the NetCDF output files generated by the previous LumiWork module, and processes them following a one-dimensional configuration-coordinate model [@jia2017first;@bouquiaux2021importance]. This analysis provides detailed insights into the luminescence characteristics of the defect under study by computing properties such as transition energies, Huang-Rhys factors, effective phonon frequencies, and lineshapes following this 1D model or within a semi-classical approximation.
[…]
