+++ Lumabi_JOSS/paper.md
@@ -52,6 +52,7 @@ Lumabi is a Python package within the AbiPy framework [@gonze2020abinit] that au
 
 Lumabi addresses the growing need for reproducible, automated workflows in defect physics [@lejaeghere2016reproducibility;@bosoni2024verify], with applications ranging from quantum technologies [@wolfowicz2021quantum;@dreyer2018first] to phosphors for solid-state lighting [@pust2015revolution;@lin2017inorganic;@fang2022evolutionary]. Tutorials and examples are available in the [AbiPy Book](https://abinit.github.io/abipy_book/lumabi/intro/intro.html).
 
+
 # Statement of need
 
 Defect-induced luminescence plays a key role in materials design for optoelectronics, quantum information, and phosphor technologies. Accurate predictions require ground- and excited-state calculations, phonon computations, and multiple post-processing steps, which are typically laborious to set up.  
@@ -64,57 +65,37 @@ Lumabi aims at filling this gap. Built on ABINIT and AbiPy, with interfaces to P
 
 The code is organized into four modules that can be combined into a complete workflow or used independently. We describe here the overall working principles of each module. For a more practical approach, we provide [online tutorials](https://abinit.github.io/abipy_book/lumabi/lumiwork/lesson_lumiwork.html).
 
+
 ## LumiWork Module
 
-![The LumiWork module, an AbiPy Workflow that automates ABINIT DFT tasks with $\Delta$SCF constrained occupations.[]{label="fig-lumiwork"}](LumiWork.pdf)
+![The LumiWork module, an AbiPy Workflow that automates ABINIT DFT tasks with $\Delta$SCF constrained occupations.](LumiWork.pdf)
 
-A computational workflow for calculating phonon-resolved photoluminescence (PL) spectra of defect systems starts with the LumiWork module (\autoref{fig-lumiwork}), which automates ABINIT DFT tasks with $\Delta$SCF constrained occupations [@jones1989density;@hellman2004potential]. Users provide the defect supercell structure (e.g. in CIF format), the DFT input parameters (e.g., the plane-wave kinetic energy cutoff, the stopping SCF criterion, etc.), and constrained occupations of the Kohn-Sham states designed to mimic the excited state of the system under study. This module manages two structural relaxations for the ground- and the excited-state, and offers optional static SCF computations followed by non-SCF band structure calculations. Note that the workflow is dynamic: since the relaxed excited state is not known beforehand, input files must be generated at run-time.
+A computational workflow for calculating phonon-resolved photoluminescence (PL) spectra of defect systems starts with the LumiWork module (Figure 2), which automates ABINIT DFT tasks with $\Delta$SCF constrained occupations [@jones1989density;@hellman2004potential]. Users provide the defect supercell structure, the DFT input parameters, and constrained occupations of the Kohn-Sham states designed to mimic the excited state of the system under study. This module manages two structural relaxations for
+the ground- and the excited-state, and offers optional static SCF computations followed by non-SCF band structure calculations. As the relaxed excited state is not known in advance, input files are generated dynamically.
 
-The `LumiWork` class inherits from the Abipy's `Work` class.
-Consequently, it leverages error detection and automatic restart handling capabilities available for all ABINIT-related Works,
-ensuring that most computations proceed smoothly without interruption.
-The majority of outputs are in NetCDF format, serving as the foundation for subsequent modules
 
 ## $\Delta$SCF Post-Processing Module
 
-![The $\Delta$SCF module, designed to post-process $\Delta$SCF constrained-occupation calculations using a one-dimensional configuration-coordinate model.[]{label="fig-post_process"}](dSCF_post_process.pdf)
[…]
