### 2026-02-27 #9145 @jbouquiaux
Hi @espottesmith ,
Just to notify you that I've modified the paper.md to address your three points.
