### 2026-02-27 #9145 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
