### 2026-03-13 62e7fe0
Add Sigres.write_qpdata method
