### 2026-03-15 #9145 @Kevin-Mattheus-Moerman
@jbouquiaux As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

Your license file is not a standard license file. It looks like the text you have in the license file describes what license you use but does not contain the license itself. The text you have there may need to move to the README or be removed. Instead please can you work to have the LICENSE text conform to the standard that the OSI recommend, i.e. to use exactly the license text as used here: https://opensource.org/license/gpl-2.0 (I assumed this is the license you intend to use).

👉 Can you update the citation to Lee et al.. It looks like you used to refer to an Arxiv reference and now refer to the published paper DOI. However the text still says `arXiv Preprint arXiv:2502.18859`, perhaps you can update the citation date to be for the final paper fully.
- [x] Copyedit pass/checked for typos
