+++ Lumabi_JOSS/paper.bib
@@ -268,12 +268,25 @@
  year = {2021}
 }
 
-@article{lee2025machine,
- author = {Lee, Nakyung and S{\'o}jka, Ma{\l}gorzata and La, Annie and Sharma, Syna and Kavanagh, Se{\'a}n and Ahn, Docheon and Scanlon, David O and Brgoch, Jakoah},
- doi = {10.1016/j.materresbull.2012.03.017},
- journal = {arXiv preprint arXiv:2502.18859},
- title = {Machine Learning a Phosphor's Excitation Band Position},
- year = {2025}
+@misc{lee2025machine,
+      title={Machine Learning a Phosphor's Excitation Band Position}, 
+      author={Nakyung Lee and Małgorzata Sójka and Annie La and Syna Sharma and Seán Kavanagh and Docheon Ahn and David O. Scanlon and Jakoah Brgoch},
+      year={2025},
+      eprint={2502.18859},
+      archivePrefix={arXiv},
+      primaryClass={cond-mat.mtrl-sci},
+      url={https://arxiv.org/abs/2502.18859}, 
+}
+
+@article{kumar2025machine,
+  title={Machine-Learning-Assisted Discovery of Cr3+-Based Near-Infrared Phosphors},
+  author={Kumar, Amit and Akbar, Arslan and Lesmes, Hannah and Kavanagh, Se{\'a}n R and Scanlon, David O and Brgoch, Jakoah},
+  journal={Chemistry of Materials},
+  volume={37},
+  number={19},
+  pages={7762--7770},
+  year={2025},
+  publisher={ACS Publications}
 }
 
 @article{lejaeghere2016reproducibility,
