### 2026-03-17 #9145 @jbouquiaux
Hi @Kevin-Mattheus-Moerman, I'm working on it. I'm in contact with @gmatteo for the License text.
