### 2026-03-18 0e9f3d2
Add parameter for interpolating QP corrections using pade
