### 2026-03-19 #9145 @Kevin-Mattheus-Moerman
@jbouquiaux on the license text. I think  you should just do the following:

1. Copy exactly this text: https://opensource.org/license/gpl-2.0 and have that be the sole content of the LICENSE file you have.
2. Next if you read the section stating `How to Apply These Terms to Your New Programs` you can see that they suggest some text to add to individual code files if you want ("...attach them to the start of each source file..."). I think the problem is that you are currently using that text as the license, but it is not the license, it is a suggested boilerplate text to refer to it at the end of your code files. So if you want to do this then do add this text to selected files.
3. Finally, once the LICENSE file is correct, you can refer to (and link to) it also in the README.
