### 2026-03-24 967a6dd
Update license information in README
