+++ README.rst
@@ -482,8 +482,17 @@ To contribute to Abipy, the standard procedure is as follows:
 License
 =======
 
-AbiPy is released under the GNU GPL license. For more details see the LICENSE file.
+Abipy is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License version 2 as published by
+the Free Software Foundation
 
+Abipy is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along with abipy.  
+If not, see http://www.gnu.org/licenses/.
 .. _Python: http://www.python.org/
 .. _Abinit: https://www.abinit.org
 .. _abinit-channel: https://anaconda.org/abinit
