### 2026-03-24 aa37b71
Fix formatting issues in README.rst
