+++ README.rst
@@ -493,6 +493,7 @@ GNU Lesser General Public License for more details.
 
 You should have received a copy of the GNU Lesser General Public License along with abipy.  
 If not, see http://www.gnu.org/licenses/.
+
 .. _Python: http://www.python.org/
 .. _Abinit: https://www.abinit.org
 .. _abinit-channel: https://anaconda.org/abinit
