### 2026-03-30 #9145 @jbouquiaux
Hi @Kevin-Mattheus-Moerman, thanks for the message.
A PR was created with the requested changes https://github.com/abinit/abipy/pull/392. I don't have the rights to merge myself. Hopefully this should be done soon. @gmatteo, do you have any update?
