### 2026-03-30 #9145 @gmatteo
@jbouquiaux,  @Kevin-Mattheus-Moerman

The pull request abinit/abipy#392 has been merged.
