+++ Lumabi_JOSS/paper.bib
@@ -280,6 +280,7 @@
 
 @article{kumar2025machine,
   title={Machine-Learning-Assisted Discovery of Cr3+-Based Near-Infrared Phosphors},
+  doi={10.1021/acs.chemmater.5c01208},
   author={Kumar, Amit and Akbar, Arslan and Lesmes, Hannah and Kavanagh, Se{\'a}n R and Scanlon, David O and Brgoch, Jakoah},
   journal={Chemistry of Materials},
   volume={37},
