### 2026-03-31 #9145 @jbouquiaux
Sorry @gmatteo, the reference was not formatted properly, my bad. It should be fine after this is merged: https://github.com/abinit/abipy/pull/393
