### 2026-04-15 #9145 @jbouquiaux
Hi @Kevin-Mattheus-Moerman, the last steps have been finalized, sorry for the delay.
- [License text](https://github.com/abinit/abipy/blob/lumabi_JOSS/LICENSE) now contains the license itself.
- References have been updated.
